Add files via upload
A demo(very limited features for now) Using Python to implement a simple printer function. The code will pull all printers in the system, and present a list with all printer names in a simple user UI and ask the user to choose the printer. The UI intended to offer simple config such as 1 or 2 side printing, etc. The code should work at least on Win, MacOS, and ChromeOS.